requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
numpy
keybert
faiss-cpu
sentence-transformers